Simplified what is in the __init__
